moved project code into early_stopping directory
