+++ early_stopping/strategies/AbstractPatienceStrategy.py
