+++ early_stopping/strategies/AbstractStrategy.py
