+++ early_stopping/strategies/LinearAdaptivePatienceStrategy.py
