+++ early_stopping/strategies/PolynomialAdaptivePatienceStrategy.py
