+++ early_stopping/strategies/SimplePatienceStrategy.py
