+++ early_stopping/strategies/SlidingWindowStrategy.py
