+++ early_stopping/strategies/StrategyFactory.py
